refactory_ebpf: Fix the MinMax 2.5 fixes

+++ src/collectors/ebpf.plugin/ebpf_dcstat.c
@@ -261,7 +261,7 @@ static inline int ebpf_dc_load_and_attach(struct dc_bpf *obj, ebpf_module_t *em)
  */
 void dcstat_update_publish(netdata_publish_dcstat_t *out, uint64_t cache_access, uint64_t not_found)
 {
-    NETDATA_DOUBLE successful_access = (NETDATA_DOUBLE)(cache_access - not_found);
+    NETDATA_DOUBLE successful_access = (NETDATA_DOUBLE)(((long long)cache_access) - ((long long)not_found));
     if (successful_access < 0)
         successful_access = 0;
     NETDATA_DOUBLE ratio = (cache_access) ? successful_access / (NETDATA_DOUBLE)cache_access : 0;
